pyproject.toml
[project]
name = "vivaestate-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "olamide",email = "olamoyegunolamidesammy@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-cors-headers (>=4.6.0,<5.0.0)",
    "django-tenants (>=3.7.0,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "drf-nested-routers (>=0.94.1,<0.95.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "django-filter (>=24.3,<25.0)",
    "djaodjin-saas (>=1.0.0,<2.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "celery (>=5.4.0,<6.0.0)",
    "twilio (>=9.4.4,<10.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "phonenumbers (>=8.13.54,<9.0.0)",
    "drf-yasg (>=1.21.8,<2.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "drf-spectacular-sidecar (>=2025.3.1,<2026.0.0)",
    "openai (>=1.65.5,<2.0.0)",
    "numpy (>=2.2.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
js2py = "^0.74"

